backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain-google-genai==0.0.6
mcp-use>=0.1.0
pydantic>=2.6.0